requirements.txt
Django>=1.4
django-classy-tags>=0.3.4.1
